Updated references to swiftype
