+++ Gemfile
@@ -1,4 +1,4 @@
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in swiftype-rb.gemspec
+# Specify your gem's dependencies in elastic-enterprise-search.gemspec
 gemspec
