+++ NOTICE.txt
@@ -1,0 +1,3 @@
+Elastic Enterprise Search Ruby client.
+
+ Copyright 2012-2019 Elasticsearch B.V.
