+++ README.md
@@ -1,6 +1,6 @@
-<p align="center"><img src="https://github.com/swiftype/swiftype-enterprise-ruby/blob/master/logo-enterprise-search.png?raw=true" alt="Elastic Enterprise Search Logo"></p>
+<p align="center"><img src="https://github.com/elastic/enterprise-search-ruby/blob/master/logo-enterprise-search.png?raw=true" alt="Elastic Enterprise Search Logo"></p>
 
-<a href="https://github.com/swiftype/swiftype-enterprise-ruby/releases"><img src="https://img.shields.io/github/release/swiftype/swiftype-enterprise-ruby/all.svg?style=flat-square" alt="GitHub release" /></a></p>
+<a href="https://github.com/elastic/enterprise-search-ruby/releases"><img src="https://img.shields.io/github/release/elastic/enterprise-search-ruby/all.svg?style=flat-square" alt="GitHub release" /></a></p>
 
 > A first-party Ruby client for [Elastic Enterprise Search](https://www.elastic.co/solutions/enterprise-search).
 
@@ -19,23 +19,23 @@
 To install the gem, execute:
 
 ```bash
-gem install swiftype-enterprise
+gem install elastic-enterprise-search
 ```
 
-Or place `gem 'swiftype-enterprise', '~> 1.0.0` in your `Gemfile` and run `bundle install`.
+Or place `gem 'elastic-enterprise-search', '~> 0.1.` in your `Gemfile` and run `bundle install`.
 
 ## Usage
 
-Create a new instance of the Swiftype Enterprise Client with your access token:
+Create a new instance of the Enterprise Search Client with your access token:
 
-    SwiftypeEnterprise.access_token = '' # your access token
-    swiftype = SwiftypeEnterprise::Client.new
+    Elastic::EnterpriseSearch.access_token = '' # your access token
+    client = Elastic::EnterpriseSearch::Client.new
 
 ### Change API endpoint
 
  ```
-swiftype = SwiftypeEnterprise::Client.new
-SwiftypeEnterprise.endpoint = 'https://your-server.example.com/api/v1'
+client = Elastic::EnterpriseSearch::Client.new
+Elastic::EnterpriseSearch.endpoint = 'https://your-server.example.com/api/v1'
 ```
 
 
@@ -60,9 +60,9 @@ This example shows how to use the index_documents method:
     ]
 
     begin
-      document_receipts = swiftype.index_documents(content_source_key, documents)
+      document_receipts = client.index_documents(content_source_key, documents)
       # handle results
-    rescue SwiftypeEnterprise::ClientException => e
+    rescue Elastic::EnterpriseSearch::ClientException => e
       # handle error
     end
 
@@ -72,9 +72,9 @@ This example shows how to use the index_documents method:
     document_ids = ['INscMGmhmX4', 'JNDFojsd02']
 
     begin
-      destroy_document_results = swiftype.destroy_documents(content_source_key, document_ids)
+      destroy_document_results = client.destroy_documents(content_source_key, document_ids)
       # handle destroy document results
-    rescue SwiftypeEnterprise::ClientException => e
+    rescue Elastic::EnterpriseSearch::ClientException => e
       # handle error
     end
 
@@ -90,17 +90,17 @@ $ ENDPOINT=http://localhost:3002/api/v1 bundle exec rspec
 
 ### Where do I report issues with the client?
 
-If something is not working as expected, please open an [issue](https://github.com/swiftype/swiftype-enterprise-ruby/issues/new).
+If something is not working as expected, please open an [issue](https://github.com/elastic/enterprise-search-ruby/issues/new).
 
 ## Contribute 🚀
 
 We welcome contributors to the project. Before you begin, a couple notes...
 
-+ Before opening a pull request, please create an issue to [discuss the scope of your proposal](https://github.com/swiftype/swiftype-enterprise-ruby/issues).
++ Before opening a pull request, please create an issue to [discuss the scope of your proposal](https://github.com/elastic/enterprise-search-ruby/issues).
 + Please write simple code and concise documentation, when appropriate.
 
 ## License 📗
 
-[MIT](https://github.com/swiftype/swiftype-enterprise-ruby/blob/master/LICENSE) © [Elastic](https://github.com/elastic)
+[Apache 2.0](https://github.com/elastic/enterprise-search-ruby/blob/master/LICENSE.txt) © [Elastic](https://github.com/elastic)
 
-Thank you to all the [contributors](https://github.com/swiftype/swiftype-enterprise-ruby/graphs/contributors)!
+Thank you to all the [contributors](https://github.com/elastic/enterprise-search-ruby/graphs/contributors)!
