+++ lib/elastic/enterprise-search.rb
@@ -1,0 +1,7 @@
+require 'elastic/enterprise-search/client'
+
+module Elastic
+  module EnterpriseSearch
+    extend Elastic::EnterpriseSearch::Configuration
+  end
+end
