+++ lib/elastic/enterprise-search/utils.rb
@@ -1,0 +1,15 @@
+module Elastic
+  module EnterpriseSearch
+    module Utils
+      extend self
+
+      def stringify_keys(hash)
+        output = {}
+        hash.each do |key, value|
+          output[key.to_s] = value
+        end
+        output
+      end
+    end
+  end
+end
