+++ lib/elastic/enterprise-search/version.rb
@@ -1,0 +1,5 @@
+module Elastic
+  module EnterpriseSearch
+    VERSION = '3.0.0'
+  end
+end
