--- lib/swiftype-enterprise.rb
@@ -1,6 +1,0 @@
-require 'swiftype-enterprise/client'
-
-module SwiftypeEnterprise
-  extend SwiftypeEnterprise::Configuration
-end
-
