--- lib/swiftype-enterprise/exceptions.rb
@@ -1,9 +1,0 @@
-module SwiftypeEnterprise
-  class ClientException < StandardError; end
-  class NonExistentRecord < ClientException; end
-  class InvalidCredentials < ClientException; end
-  class BadRequest < ClientException; end
-  class Forbidden < ClientException; end
-  class UnexpectedHTTPException < ClientException; end
-  class InvalidDocument < ClientException; end
-end
