--- lib/swiftype-enterprise/version.rb
@@ -1,3 +1,0 @@
-module SwiftypeEnterprise
-  VERSION = '3.0.0'
-end
